+++ variant-switcher/src/shared.ts
@@ -1,6 +1,16 @@
+/**
+ Copyright (c) 2021-present, Eaton
+ All rights reserved.
+ This code is licensed under the BSD-3 license found in the LICENSE file in the root directory of this source tree and at https://opensource.org/licenses/BSD-3-Clause.
+ **/
+
 export const KEYS = {
     PROPERTY_NAME: 'property-name',
     FROM_VARIANT: 'from-variant',
     TO_VARIANT: 'to-variant',
     DEEP_SWITCH: 'deep-switch',
+    FULL_DOCUMENT: 'full-document',
+    EXACT_MATCH: 'exact-match',
+    MAIN_COMPONENT_NAME: 'main-component-name',
+    SHOW_ADVANCED_OPTIONS: 'show-advanced-options',
 };